add Group component, refractor TaskGroups
